refactor(tui): split lib.rs into input_translate and review_flow (ADR 0028)
lib.rs had grown to 1389 lines (past the warn threshold). Split along
responsibility, not line count:

- input_translate.rs: translate_key, translate_mouse_event, and
  normalize_fullwidth_key — raw crossterm input to InputKey, no App
  mutation.
- review_flow.rs: dispatch_note_compose_key, derive_selection_snapshot,
  first_anchor_run, perform_export, clipboard_export_status, and
  should_recompute_note_markers — the ADR 0048 review-notes
  composing/exporting/caching glue run_app's event loop delegates to.

lib.rs itself drops to 925 lines; run_app's own dispatch machinery
(dispatch_non_source_key, the diff-pane/blast-radius recompute gates,
apply_diff_pane_selection_effects, resolve_goto) stays there since it
is one cohesive event-loop responsibility. No behavior change; all 807
tests pass unmodified except for re-exporting two review_flow items at
the crate root under #[cfg(test)] so the existing lib_tests imports
keep working.
